backend/requirements.txt
langchain-community: 0.2.7
langchain-openai: 0.1.16
langchain-core: 0.2.17
langchain: 0.2.7
openai: 1.35.13
python-dotenv: 1.0.1
faiss-cpu: 1.8.0.post1
boto3: 1.34.144
pypdf: 4.2.0
Flask: 3.0.3
Flask-Cors: 4.0.1
requests: 2.32.3
PyJWT: 2.3.0
cryptography: 42.0.8
redis: 5.0.7
flask-redis: 0.4.0
